fix: take full ownership of atem media uploading, and remove hook for call from core. Trigger check even if device is warning
